Push sprite building into a worker
